Get rid of signed/unsigned warning
[git-p4: depot-paths = "//depot/qt/main/": change = 33589]

+++ src/kernel/qurl.cpp
@@ -75,7 +75,7 @@ static void slashify( QString& s, bool allowMultiple = TRUE )
 	}
 	if ( s[ i ] == '\\' )
 	    s[ i ] = '/';	    
-	if ( s[ i ] == ':' && (i == s.length()-1 || s[ i + 1 ] != '/' ) ) //mac colon's go away, unless after a protocol
+	if ( s[ i ] == ':' && (i == (int)s.length()-1 || s[ i + 1 ] != '/' ) ) //mac colon's go away, unless after a protocol
 		s[ i ] = '/';	
 	if ( s[ i ] == '/' )
 	    justHadSlash = TRUE;
